service: рефакторит favorites-screen и компоненты
